feat: cluster event handler props at bottom of property pane
